Mainsite should now validate as HTML5.
Going to start changing the HTML syntax to match the new HTML5 tags in a new branch in my local machine. I can share it remotely if you want.

(Don't worry, there's a small hack that makes this syntax work in IE. No video or anything like that though - but who needs it?)